refactor: 경주 기록 관리 새로운 클래스로 분리
